requirements.txt
tkinter
Pillow
numpy
pymatgen
